saving - might need to rework mem-functions

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: dulrich <dulrich@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 12:22:44 by dulrich           #+#    #+#             */
-/*   Updated: 2023/09/14 11:33:43 by dulrich          ###   ########.fr       */
+/*   Updated: 2023/09/14 13:10:43 by dulrich          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] == s2[i] && n > i)
